+++ src/store/actions/centeringModesActions/centeringModesActions.ts
@@ -1,0 +1,8 @@
+import { CenteringMode } from 'enums/CenteringMode';
+
+export const SET_ACTIVE_CENTERING_MODE = 'SET_ACTIVE_CENTERING_MODE';
+
+export const setActiveCenteringMode = (payload: CenteringMode) => ({
+  type: SET_ACTIVE_CENTERING_MODE,
+  payload,
+});
