+++ src/store/actions/centeringModesActions/index.ts
@@ -1,8 +1,1 @@
-import { CenteringMode } from 'enums/CenteringMode';
-
-export const SET_ACTIVE_CENTERING_MODE = 'SET_ACTIVE_CENTERING_MODE';
-
-export const setActiveCenteringMode = (payload: CenteringMode) => ({
-  type: SET_ACTIVE_CENTERING_MODE,
-  payload,
-});
+export * from './centeringModesActions';
