+++ src/store/actions/customizationActions/customizationActions.ts
@@ -1,0 +1,5 @@
+export const TOGGLE_CUSTOMIZATION = 'TOGGLE_CUSTOMIZATION';
+
+export const toggleCustomization = () => ({
+  type: TOGGLE_CUSTOMIZATION,
+});
