+++ src/store/actions/customizationActions/index.ts
@@ -1,5 +1,1 @@
-export const TOGGLE_CUSTOMIZATION = 'TOGGLE_CUSTOMIZATION';
-
-export const toggleCustomization = () => ({
-  type: TOGGLE_CUSTOMIZATION,
-});
+export * from './customizationActions';
