+++ src/store/actions/mapActions/index.ts
@@ -1,14 +1,1 @@
-import type { Coords } from 'types/Coords';
-
-export const UPDATE_COORDS = 'UPDATE_COORDS';
-export const UPDATE_ZOOM_LEVEL = 'UPDATE_ZOOM_LEVEL';
-
-export const updateCoords = (payload: Coords) => ({
-  type: UPDATE_COORDS,
-  payload,
-});
-
-export const updateZoomLevel = (payload: number) => ({
-  type: UPDATE_ZOOM_LEVEL,
-  payload,
-});
+export * from './mapActions';
