+++ src/store/actions/themeActions/index.ts
@@ -1,8 +1,1 @@
-import { Theme } from 'enums/Theme';
-
-export const SET_ACTIVE_THEME = 'SET_ACTIVE_THEME';
-
-export const setActiveTheme = (payload: Theme) => ({
-  type: SET_ACTIVE_THEME,
-  payload,
-});
+export * from './themeActions';
