+++ src/store/reducers/customizationReducer/index.ts
@@ -1,25 +1,1 @@
-import { TOGGLE_CUSTOMIZATION } from '../../actions';
-
-import { CustomizationState } from 'interfaces/CustomizationState';
-
-import type { CustomizationActionTypes } from 'types/CustomizationActionTypes';
-
-const initialState: CustomizationState = {
-  isCustomizationEnabled: true,
-};
-
-export const customizationReducer = (
-  state = initialState,
-  action: CustomizationActionTypes
-) => {
-  switch (action.type) {
-    case TOGGLE_CUSTOMIZATION:
-      return {
-        ...state,
-        isCustomizationEnabled: !state.isCustomizationEnabled,
-      };
-
-    default:
-      return state;
-  }
-};
+export { customizationReducer } from './customizationReducer';
