+++ src/store/reducers/searchHistoryReducer/searchHistoryReducer.ts
@@ -1,0 +1,42 @@
+import {
+  SEARCH_HISTORY_ADD,
+  SEARCH_HISTORY_REMOVE,
+  SEARCH_HISTORY_CLEAR,
+} from '../../actions';
+
+import { SearchHistoryState } from 'interfaces/SearchHistoryState';
+
+import type { SearchHistoryActionTypes } from 'types/SearchHistoryActionTypes';
+
+const initialState: SearchHistoryState = {
+  items: [],
+};
+
+export const searchHistoryReducer = (
+  state = initialState,
+  action: SearchHistoryActionTypes
+) => {
+  switch (action.type) {
+    case SEARCH_HISTORY_ADD: {
+      const exists = state.items.find(
+        (item) => item.place_id === action.payload.place_id
+      );
+
+      if (exists) return state;
+
+      return { ...state, items: [action.payload, ...state.items] };
+    }
+
+    case SEARCH_HISTORY_REMOVE:
+      return {
+        ...state,
+        items: state.items.filter((item) => item.place_id !== action.payload),
+      };
+
+    case SEARCH_HISTORY_CLEAR:
+      return { ...state, items: initialState.items };
+
+    default:
+      return state;
+  }
+};
