+++ src/store/reducers/themeReducer/themeReducer.ts
@@ -1,0 +1,24 @@
+import { SET_ACTIVE_THEME } from '../../actions';
+
+import { Theme } from 'enums/Theme';
+
+import { ThemeState } from 'interfaces/ThemeState';
+
+import type { ThemeActionTypes } from 'types/ThemeActionTypes';
+
+const initialState: ThemeState = {
+  activeTheme: Theme.Light,
+};
+
+export const themeReducer = (
+  state = initialState,
+  action: ThemeActionTypes
+) => {
+  switch (action.type) {
+    case SET_ACTIVE_THEME:
+      return { ...state, activeTheme: action.payload };
+
+    default:
+      return state;
+  }
+};
